requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
# PostgreSQL drivers - BSD-licensed alternatives
pg8000==1.30.1  # BSD license - pure Python PostgreSQL driver
# psycopg2-binary==2.9.9  # LGPL - REMOVED for license compatibility
geoalchemy2==0.14.2

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Monitoring & Logging
structlog==23.2.0

# Spatial data processing
shapely==2.0.2
geojson==3.1.0
